fix: resolve all real clang-tidy findings across entire codebase
- readability-math-missing-parentheses: add explicit parentheses at 39
  sites across 13 files to clarify operator precedence
- performance-for-range-copy / unnecessary-copy-initialization: 14
  range-for loop variables and copy-inits changed to const refs
- bugprone-crtp-constructor-accessibility: 5 CRTP mixin bases in
  strongtype.h given protected default constructors
- bugprone-unhandled-self-assignment: history.h HistMessage::operator=
  now guards against self-assignment
- bugprone-branch-clone: collapsed 6 switch/if duplicate branches in
  iaudiosink.h, filetransferwidget.cpp, toxcoreerrorparser.cpp,
  network.cpp, history.cpp, chatformheader.cpp
- cppcoreguidelines-init-variables / pro-type-member-init: ~180
  uninitialized locals and member variables given zero initializers
  across audio, chatlog, core, ipc, model, net, persistence, platform,
  video, widget
- performance-unnecessary-value-param: ~116 function parameters changed
  from by-value to const-ref for non-trivial types (QString, QByteArray,
  ToxFile, shared_ptr, etc.) with matching header updates
- bugprone-narrowing-conversions: explicit static_cast added at ~266
  sites (qsizetype->int, size_t->ptrdiff_t, qreal->float, etc.)
- bugprone-implicit-widening-of-multiplication-result: pre-cast operands
  before multiply where result is stored in a wider type
- cppcoreguidelines-pro-type-static-cast-downcast: static_cast downcasts
  replaced with qobject_cast (QObject hierarchy) or dynamic_cast
  (QGraphicsItem hierarchy) throughout widget and chatlog code

All 35 ASan/UBSan tests pass.

+++ src/core/toxfileprogress.cpp
@@ -82,7 +82,7 @@ double ToxFileProgress::getProgress() const
     if (filesize == 0) {
         return 0.0;
     }
-    return double(samples[activeSample].bytesSent) / filesize;
+    return static_cast<double>(samples[activeSample].bytesSent) / static_cast<double>(filesize);
 }
 
 double ToxFileProgress::getSpeed() const
@@ -104,8 +104,8 @@ double ToxFileProgress::getSpeed() const
     const auto& active = samples[activeSample];
     const auto& inactive = samples[activeSample == 0 ? 1 : 0];
 
-    return (active.bytesSent - inactive.bytesSent)
-           / double(inactive.timestamp.msecsTo(active.timestamp)) * 1000.0;
+    return static_cast<double>(active.bytesSent - inactive.bytesSent)
+           / static_cast<double>(inactive.timestamp.msecsTo(active.timestamp)) * 1000.0;
 }
 
 double ToxFileProgress::getTimeLeftSeconds() const
